Rename mixin module
